refactor: move QualitySettings to its own file
