+++ react-components/src/architecture/base/concreteCommands/SetQualityCommand.ts
@@ -3,9 +3,9 @@
  */
 
 import { type TranslationInput } from '../utilities/TranslateInput';
-import { type QualitySettings } from '../../../components/RevealToolbar/SettingsContainer/types';
 import { type DataSourceType, type Cognite3DViewer } from '@cognite/reveal';
 import { RenderTargetCommand } from '../commands/RenderTargetCommand';
+import { QualitySettings } from '../utilities/quality/QualitySettings';
 
 export class SetQualityCommand extends RenderTargetCommand {
   // ==================================================
