+++ react-components/src/architecture/base/utilities/quality/QualitySettings.ts
@@ -1,0 +1,7 @@
+import type { CadModelBudget, PointCloudBudget, ResolutionOptions } from '@cognite/reveal';
+
+export type QualitySettings = {
+  cadBudget: CadModelBudget;
+  pointCloudBudget: PointCloudBudget;
+  resolutionOptions: ResolutionOptions;
+};
