+++ react-components/src/components/RevealToolbar/RevealToolbar.tsx
@@ -11,7 +11,7 @@ import { withSuppressRevealEvents } from '../../higher-order-components/withSupp
 import { HelpButton } from './HelpButton';
 import { ShareButton } from './ShareButton';
 import { ResetCameraButton } from './ResetCameraButton';
-import { type QualitySettings } from './SettingsContainer/types';
+import { type QualitySettings } from '../../architecture/base/utilities/quality/QualitySettings';
 import styled from 'styled-components';
 import { SelectSceneButton } from './SelectSceneButton';
 import { RuleBasedOutputsButton } from './RuleBasedOutputsButton';
