+++ react-components/src/components/RevealToolbar/SettingsContainer/types.ts
@@ -2,12 +2,8 @@
  * Copyright 2023 Cognite AS
  */
 
-import {
-  type CadModelBudget,
-  type PointCloudBudget,
-  type ResolutionOptions
-} from '@cognite/reveal';
 import { type ReactElement } from 'react';
+import { QualitySettings } from '../../../architecture/base/utilities/quality/QualitySettings';
 
 export type QualityProps = {
   lowQualitySettings?: Partial<QualitySettings>;
@@ -16,10 +12,4 @@ export type QualityProps = {
 
 export type SettingsContainerProps = QualityProps & {
   customSettingsContent?: ReactElement;
-};
-
-export type QualitySettings = {
-  cadBudget: CadModelBudget;
-  pointCloudBudget: PointCloudBudget;
-  resolutionOptions: ResolutionOptions;
 };
