+++ react-components/src/components/RevealToolbar/index.ts
@@ -6,4 +6,3 @@ export { RevealToolbar } from './RevealToolbar';
 export type { RevealToolbarProps } from './RevealToolbar';
 export type { LayersButtonProps } from './LayersButton/LayersButton';
 export type { LayersUrlStateParam, DefaultLayersConfiguration } from './LayersButton/types';
-export type { QualitySettings } from './SettingsContainer/types';
